Fix most float conversion warnings
Fixes ~200 out of ~240 `-Wfloat-conversion` warnings on gcc 13.2

+++ libhdf5/hdf5internal.c
@@ -995,7 +995,7 @@ nc4_hdf5_get_chunk_cache(int ncid, size_t *sizep, size_t *nelemsp,
     if (H5Pget_cache(plistid, NULL, nelemsp, sizep, &dpreemption) < 0)
 	return NC_EHDFERR;
     if (preemptionp)
-	*preemptionp = dpreemption;
+	*preemptionp = (float)dpreemption;
 
     return NC_NOERR;
 }
